+++ CHANGES.md
@@ -1,5 +1,6 @@
 ## Version dev-1.2.0
 - [FEATURE] Add StructuredDataUtility
+- [FEATURE] Add Recipe structured data for Detail view
 - [CHANGE] Recipe in List is an itemprop Article
 - [CHANGE] Use FlexForm/TypoScript media size settings